# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Finance Manager API",
    "description": "REST API for the Finance Manager personal finance tracker. All endpoints are profile-scoped via the `X-Profile-Id` header.",
    "version": "1.1.0",
    "license": { "name": "AGPL-3.0-only" },
    "contact": { "url": "https://github.com/Komediruzecki/finance-manager" }
  },
  "servers": [{ "url": "http://localhost:3847", "description": "Local development" }],
  "security": [{ "cookieAuth": [] }],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" },
          "429": { "description": "Rate limited" }
        }
      }
    },
...
  },
...
}